+++ src/data/OrgUnitD2ApiRepository.ts
@@ -6,74 +6,28 @@ class OrgUnitD2ApiRepository implements OrgUnitRepository {
     constructor(private d2Api: D2Api) {}
 
     async getByIds(ids: Id[]): Promise<OrgUnit[]> {
-        const { objects } = await this.d2Api.models.organisationUnits
-            .get({
+        const response = await this.d2Api
+            .get<{ organisationUnits: OrgUnit[] }>("/organisationUnits", {
                 paging: false,
-                fields: {
-                    id: true,
-                    featureType: true,
-                    coordinates: true,
-                },
-                filter: {
-                    id: { in: ids },
-                },
+                fields: "id,geometry",
+                filter: `id:in:[${ids.join(",")}]`,
             })
             .getData();
 
-        const orgUnitsData = objects.map(o => o as OrgUnitData);
-
-        return orgUnitsData.map(this.mapToDomain);
+        return response.organisationUnits;
     }
 
     async getAllWithCoordinates(): Promise<OrgUnit[]> {
-        //TODO: use this.d2Api.models.organisationUnits when !null operator filter
-        //may be provided
         const response = await this.d2Api
-            .get<{ organisationUnits: OrgUnitData[] }>("/organisationUnits", {
+            .get<{ organisationUnits: OrgUnit[] }>("/organisationUnits", {
                 paging: false,
-                fields: "id,featureType,coordinates",
-                filter: "coordinates:!null",
+                fields: "id,geometry",
+                filter: "geometry:!null",
             })
             .getData();
 
-        const domainOrgUnits = response.organisationUnits.map(this.mapToDomain);
-
-        const orgUnitsWithGeometry = domainOrgUnits.filter(ou => ou.geometry).map(this.mapToDomain);
-
-        return orgUnitsWithGeometry;
-    }
-
-    private mapToDomain(orgUnitsData: OrgUnitData): OrgUnit {
-        try {
-            const coordinates = orgUnitsData.coordinates
-                ? JSON.parse(orgUnitsData.coordinates)
-                : null;
-
-            switch (orgUnitsData.featureType) {
-                case "POINT":
-                    return {
-                        id: orgUnitsData.id,
-                        geometry: { type: "Point", coordinates: coordinates },
-                    };
-                case "POLYGON":
-                case "MULTI_POLYGON":
-                    return {
-                        id: orgUnitsData.id,
-                        geometry: { type: "Polygon", coordinates: coordinates },
-                    };
-                default:
-                    return { id: orgUnitsData.id };
-            }
-        } catch (error) {
-            return { id: orgUnitsData.id };
-        }
+        return response.organisationUnits;
     }
 }
 
 export default OrgUnitD2ApiRepository;
-
-export interface OrgUnitData {
-    id: string;
-    featureType?: "NONE" | "MULTI_POLYGON" | "POLYGON" | "POINT" | "SYMBOL";
-    coordinates?: string;
-}
